[ska] relocate functional platform tests (#226470)
## Summary

Part of https://github.com/elastic/kibana-team/issues/1503

This PR is mostly about moving platform tests from x-pack/test

```
x-pack/test/
     | - examples/
     | - functional_cors/
     | - functional_embedded/
```

After:
```
x-pack/platform/test/
     | - examples/
     | - functional_cors/
     | - functional_embedded/
```

---------

(cherry picked from commit ab965f53604cf5cd94a4f282c5e832964abfec39)
