+++ package.json
@@ -595,7 +595,7 @@
     "@kbn/home-sample-data-types": "link:src/platform/packages/shared/home/sample_data_types",
     "@kbn/i18n": "link:src/platform/packages/shared/kbn-i18n",
     "@kbn/i18n-react": "link:src/platform/packages/shared/kbn-i18n-react",
-    "@kbn/iframe-embedded-plugin": "link:x-pack/test/functional_embedded/plugins/iframe_embedded",
+    "@kbn/iframe-embedded-plugin": "link:x-pack/platform/test/functional_embedded/plugins/iframe_embedded",
     "@kbn/image-embeddable-plugin": "link:src/platform/plugins/private/image_embeddable",
     "@kbn/index-adapter": "link:x-pack/solutions/security/packages/index-adapter",
     "@kbn/index-lifecycle-management-common-shared": "link:x-pack/platform/packages/shared/index-lifecycle-management/index_lifecycle_management_common_shared",
@@ -630,7 +630,7 @@
     "@kbn/kbn-tp-run-pipeline-plugin": "link:src/platform/test/interpreter_functional/plugins/kbn_tp_run_pipeline",
     "@kbn/key-value-metadata-table": "link:x-pack/platform/packages/shared/kbn-key-value-metadata-table",
     "@kbn/kibana-api-cli": "link:x-pack/platform/packages/shared/kbn-kibana-api-cli",
-    "@kbn/kibana-cors-test-plugin": "link:x-pack/test/functional_cors/plugins/kibana_cors_test",
+    "@kbn/kibana-cors-test-plugin": "link:x-pack/platform/test/functional_cors/plugins/kibana_cors_test",
     "@kbn/kibana-overview-plugin": "link:src/platform/plugins/private/kibana_overview",
     "@kbn/kibana-react-plugin": "link:src/platform/plugins/shared/kibana_react",
     "@kbn/kibana-usage-collection-plugin": "link:src/platform/plugins/private/kibana_usage_collection",
