+++ tsconfig.base.json
@@ -1088,8 +1088,8 @@
       "@kbn/i18n/*": ["src/platform/packages/shared/kbn-i18n/*"],
       "@kbn/i18n-react": ["src/platform/packages/shared/kbn-i18n-react"],
       "@kbn/i18n-react/*": ["src/platform/packages/shared/kbn-i18n-react/*"],
-      "@kbn/iframe-embedded-plugin": ["x-pack/test/functional_embedded/plugins/iframe_embedded"],
-      "@kbn/iframe-embedded-plugin/*": ["x-pack/test/functional_embedded/plugins/iframe_embedded/*"],
+      "@kbn/iframe-embedded-plugin": ["x-pack/platform/test/functional_embedded/plugins/iframe_embedded"],
+      "@kbn/iframe-embedded-plugin/*": ["x-pack/platform/test/functional_embedded/plugins/iframe_embedded/*"],
       "@kbn/image-embeddable-plugin": ["src/platform/plugins/private/image_embeddable"],
       "@kbn/image-embeddable-plugin/*": ["src/platform/plugins/private/image_embeddable/*"],
       "@kbn/import-locator": ["packages/kbn-import-locator"],
@@ -1172,8 +1172,8 @@
       "@kbn/key-value-metadata-table/*": ["x-pack/platform/packages/shared/kbn-key-value-metadata-table/*"],
       "@kbn/kibana-api-cli": ["x-pack/platform/packages/shared/kbn-kibana-api-cli"],
       "@kbn/kibana-api-cli/*": ["x-pack/platform/packages/shared/kbn-kibana-api-cli/*"],
-      "@kbn/kibana-cors-test-plugin": ["x-pack/test/functional_cors/plugins/kibana_cors_test"],
-      "@kbn/kibana-cors-test-plugin/*": ["x-pack/test/functional_cors/plugins/kibana_cors_test/*"],
+      "@kbn/kibana-cors-test-plugin": ["x-pack/platform/test/functional_cors/plugins/kibana_cors_test"],
+      "@kbn/kibana-cors-test-plugin/*": ["x-pack/platform/test/functional_cors/plugins/kibana_cors_test/*"],
       "@kbn/kibana-manifest-schema": ["packages/kbn-kibana-manifest-schema"],
       "@kbn/kibana-manifest-schema/*": ["packages/kbn-kibana-manifest-schema/*"],
       "@kbn/kibana-overview-plugin": ["src/platform/plugins/private/kibana_overview"],
