+++ x-pack/platform/test/examples/embedded_lens/embedded_example.ts
