+++ x-pack/platform/test/examples/embedded_lens/index.ts
