+++ x-pack/platform/test/examples/screenshotting/index.ts
