+++ x-pack/platform/test/examples/search_examples/index.ts
