+++ x-pack/platform/test/examples/search_examples/partial_results_example.ts
