+++ x-pack/platform/test/examples/search_examples/search_example.ts
