+++ x-pack/platform/test/examples/search_examples/search_session_example.ts
