+++ x-pack/platform/test/examples/search_examples/search_sessions_cache.ts
