+++ x-pack/platform/test/examples/search_examples/sql_search_example.ts
