+++ x-pack/platform/test/examples/triggers_actions_ui_examples/global_rule_event_log_list.ts
