+++ x-pack/platform/test/examples/triggers_actions_ui_examples/rule_event_log_list.ts
