+++ x-pack/platform/test/examples/triggers_actions_ui_examples/rule_status_filter.ts
