+++ x-pack/platform/test/examples/triggers_actions_ui_examples/rules_settings_link.ts
