+++ x-pack/platform/test/functional_cors/ftr_provider_context.d.ts
