+++ x-pack/platform/test/functional_cors/plugins/kibana_cors_test/kibana.jsonc
