+++ x-pack/platform/test/functional_cors/plugins/kibana_cors_test/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../../../tsconfig.base.json",
+  "extends": "../../../../../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "target/types"
   },
@@ -11,6 +11,8 @@
     "target/**/*",
   ],
   "kbn_references": [
-    "@kbn/core"
+    "@kbn/core",
+    "@kbn/test",
+    "@kbn/config-schema",
   ]
 }
