+++ x-pack/platform/test/functional_cors/services.ts
