+++ x-pack/platform/test/functional_cors/tests/cors.ts
