+++ x-pack/platform/test/functional_cors/tests/index.ts
