+++ x-pack/platform/test/functional_embedded/ftr_provider_context.d.ts
