+++ x-pack/platform/test/functional_embedded/plugins/iframe_embedded/server/plugin.ts
