+++ x-pack/platform/test/functional_embedded/plugins/iframe_embedded/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../../../tsconfig.base.json",
+  "extends": "../../../../../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "target/types"
   },
@@ -11,8 +11,6 @@
     "target/**/*",
   ],
   "kbn_references": [
-    "@kbn/core",
-    "@kbn/test",
-    "@kbn/config-schema",
+    "@kbn/core"
   ]
 }
